requirements.txt
# --- Core Dependencies ---
opencv-python
numpy
mss
ultralytics
PyYAML
easyocr
requests
onnx

# --- PyTorch Installation ---
# Note: Choose ONE based on your hardware.

# 1. For CPU ONLY (Default):
# torch
# torchvision

# 2. For NVIDIA GPU (RTX 5060, etc.):
# DO NOT use 'pip install -r requirements.txt' for torch if you want GPU support.
# Instead, run the following command manually:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
